Add new fragments to the fragList in the createLetterFragments method rather than onAttachFragment.
When the orientation changes it will destroy the MainActivity and then re-create it.  However after destroying it and before re-creating it calls onAttachFragment for each fragment that was in the destroyed activity.  Since fragList is null after being destroyed it caused null reference errors when onAttachFragment was called before onCreate.